truss/templates/trtllm-whisper/requirements.txt
--extra-index-url https://pypi.nvidia.com
tensorrt_llm==0.9.0
hf_transfer
janus
kaldialign
librosa
mpi4py==3.1.4
safetensors
soundfile
tiktoken
torchaudio
async-batcher>=0.2.0
pydantic>=2.7.1